임시저장2

+++ src/main/java/com/sparta/itsmine/domain/auction/dto/AuctionRequestDto.java
@@ -5,5 +5,5 @@ import lombok.Getter;
 
 @Getter
 public class AuctionRequestDto {
-    private Long bidPrice;
+    private Integer bidPrice;
 }
