+++ src/main/java/com/sparta/itsmine/domain/auction/dto/GetAuctionByMaxedBidPriceResponseDto.java
@@ -9,10 +9,10 @@ import lombok.Setter;
 public class GetAuctionByMaxedBidPriceResponseDto {
 
     private Long productId;
-    private Long bidPrice;
+    private Integer bidPrice;
 
     @QueryProjection
-    public GetAuctionByMaxedBidPriceResponseDto(Long productId, Long bidPrice) {
+    public GetAuctionByMaxedBidPriceResponseDto(Long productId, Integer bidPrice) {
         this.productId = productId;
         this.bidPrice = bidPrice;
     }
