+++ src/main/java/com/sparta/itsmine/domain/auction/dto/GetAuctionByUserResponseDto.java
@@ -12,10 +12,10 @@ public class GetAuctionByUserResponseDto {
 
     private Long userId;
     private Long productId;
-    private Long bidPrice;
+    private Integer bidPrice;
 
     @QueryProjection
-    public GetAuctionByUserResponseDto(Long productId, Long bidPrice, Long userId) {
+    public GetAuctionByUserResponseDto(Long productId, Integer bidPrice, Long userId) {
         this.productId = productId;
         this.bidPrice = bidPrice;
         this.userId = userId;
